+++ activemodel/CHANGELOG.md
@@ -1,3 +1,30 @@
+*   Migrate methods from `ActiveRecord::AttributeMethods`, `ActiveRecord::Attributes::Read` and `ActiveRecord::Attributes::Write` to Active Model
+
+    ```ruby
+    class Topic
+      include ActiveModel::AttributeMethods
+
+      attr_accessor :title
+
+      alias_attribute :heading, :title
+    end
+
+    topic = Topic.new
+    topic[:title] = "Hello Title"
+    topic[:title] # => "Hello Title"
+
+    topic.write_attribute(:title, "Hello Title Method")
+    topic.read_attribute(:title) # => "Hello Title Method"
+
+    topic[:heading] = "Hello Heading"
+    topic[:heading] # => "Hello Heading"
+
+    topic.write_attribute(:heading, "Hello Heading Method")
+    topic.read_attribute(:heading) # => "Hello Heading Method"
+    ```
+
+    *Sean Doyle*
+
 *   Add `has_json` and `has_delegated_json` to provide schema-enforced access to JSON attributes.
 
     ```ruby